+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## Unreleased
 
 ### Added
+- resource `pullzone`: support `tls_level`;
 - resource `pullzone_shield`: support Bot Categorization;
 - resource `pullzone_ratelimit_rule`: support `condition.negated`, `limit.counter_key` and `response.action` ([#97](https://github.com/BunnyWay/terraform-provider-bunnynet/pull/97));
 - resource `pullzone_ratelimit_rule`: support `condition.variable_value_regex`;